requirements.txt
dj-database-url==0.4.0
dj-static==0.0.6
Django==1.7
django-toolbelt==0.0.1
gunicorn==19.4.5
psycopg2==2.6.1
pycrypto==2.6.1
static3==0.6.1
whitenoise==3.0
djangorestframework==3.3.3
markdown==2.6.6
django-filter==0.13.0